requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
Byaldi==0.0.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colpali_engine @ git+https://github.com/illuin-tech/colpali@db3d00ea97131241714bf046a09ee5bb653a52a5
datasets==3.2.0
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
mteb==1.6.35
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.19.1
networkx==2.8.8
ninja==1.11.1.3
numpy==1.26.4
ollama==0.4.4
openai==1.58.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
peft==0.11.1
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrec-eval-terrier==0.5.6
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
sentence-transformers==3.3.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
srsly==2.5.0
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
